feat: 使用 vnode + h函数 + render函数 实现函数式调用 comfirm 组件弹窗

+++ src/views/layout/components/header/header-search/search-history/index.vue
@@ -30,6 +30,7 @@
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs/index.js'
 
 const store = useStore()
 
@@ -37,10 +38,10 @@ const emits = defineEmits(['itemClick'])
 
 // 删除所有记录
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
-  // confirm('要删除所有历史记录吗？').then(() => {
-  //   store.commit('search/deleteAllHistory')
-  // })
+  // store.commit('search/deleteAllHistory')
+  confirm('要删除所有历史记录吗？').then(() => {
+    store.commit('search/deleteAllHistory')
+  })
 }
 
 // 删除单个记录
